Load asset list 3 target durations before ad start time
This makes sure that the asset list is loaded about the
same duration before the ad start as on other platforms.

Further, interstitials in live streams are taken into
account when 3 target durations before the live window.
This enables the ads loader to preload such ads in a
future change. This may be indicated for live streams
that should play as close to the live edge as possible.
In such a case the buffered position may reach the start
of the ad only late. Preloading could help here to
decrease the performance requirements of ad servers.

PiperOrigin-RevId: 746521261