Catch SecurityException caused by wrong implementation of Binder
The two-way binder interfaces in the session.legacy module can
throw SecurityExceptions if the remote app fails to correctly parse
the binder call. This has happened for apps where code obfuscation
wrongly changed the interface string constant of the generated
IMediaSession class.

To avoid crashing the sender app for this, we can treat these
cases as if the binder has died.

PiperOrigin-RevId: 756786968