Make `FakeMediaPeriod.getAdjustedSeekPositionUs(...)` more realistic
This is needed in a subsequent change for testing seek tolerance in
scrubbing mode.

This requires plumbing `syncSampleTimesUs` through `FakeMediaSource`
into `FakeMediaPeriod`, and so this change also adds a
`FakeMediaSource.Builder` as the telescoping constructors are getting a
bit unwieldy.

The two simplest (and most widely used) constructors are kept
un-deprecated.

PiperOrigin-RevId: 756800857