[ui-compose] Change PresentationState's dependency on Player
`PresentationState` will now take a nullable Player, which is in sync with and to accommodate PlayerSurface's similar change

`PresentationState` will now only be created once, rather than on any new Player change (as visible by removing the player from the `key` of `remember`). This helps utilize `keepContentOnReset` which was set in a previous composition, but now will be part of the same PresentationState object in between Player changes. Visually, that means if `keepContentOnReset=true`, there should be no shutter if the Player changes.

PiperOrigin-RevId: 756852419