[ui-compose] Add SeekBack and SeekForward button states
[demo-compose]
Note that the buttons are no longer called Rewind and FastForward, as they used to be in PlayerView/PlayerControlView.

The alternative to using predefined drawables was to use Material3 Icons, but they are limited to Icons.Default.Replay5, 10, 30 and an empty arrow. Same for Forward5, 10, 30 and an empty arrow. In order to support any seek back and forward increment, we would have to overlay an empty circular arrow with a Text composable which does not scale well.

PiperOrigin-RevId: 757741047