Increase the default image buffer size in `DefaultLoadControl`
The existing value of `128kB` seems absurdly small (probably copy-pasted
from text and metadata above [1]?). In many cases this will be overridden to
`DEFAULT_MIN_BUFFER_SIZE = 13.1MB`, but if we have a video track and an
image track then this override won't kick in.

By setting this to 26MB we make it large enough to handle 50MP Ultra
HDR images such as the 13.8MB example in Issue: androidx/media#2417 (images are
loaded as a single sample, they are not streamed in chunks, so the
buffer has to be large enough to hold the whole image).

[1] <unknown commit>

PiperOrigin-RevId: 758188606