DVFP: ensure pending input streams are configured on flush
flush() can be called multiple times when a pending stream is
waiting to be configured. For example, renderer position is reset
when the renderer is enabled, and when the user seeks - both of which
cause a position reset.

DVFP.flush() flushes pending tasks on the GL thread - one of these tasks
is configuring the pending stream.

Fixes a flaky CompositionPlayerSeekTest and a race condition during seeking

PiperOrigin-RevId: 759036193