Start notification manager in 'dismissed' state to wait for prepare
Currently we show a notification immediately when the player is
set up with media, even if it never starts playing. This case may
happen for pure media browsing use cases for example. To avoid
showing a notification that is never used, we can start the
notification manager in the 'dismissed' state so that we wait for
the first non-STATE_IDLE update before showing a notification.

Issue: androidx/media#2423
PiperOrigin-RevId: 759595094