Deflake ServerSideAdInsertionMediaSourceTest
Waiting for a specific playback position can be tricky if there
is an unknown current number of queued messages on the playback
thread and playback may have already moved past the desired
point in time.

In this case, this meant the seek in the test failed because
playback moved on to be inside the next midroll already.
This can be fixed by using untilBackgroundThreadCondition, which
is more robust against this kind of issue.

PiperOrigin-RevId: 760638400