Allow apps to override MediaButtonReceiver.onReceive
Apps are on their own when doing that. The main reasons to do
may be that an app needs to get information whether to start
the service or not from an async call.

The system may have stopped the app or process before the async
call returns. This behaviour is considered risky by Media3 so
app can do that but we don't provide support for that.

Issue: androidx/media#2443
PiperOrigin-RevId: 761462217