Clean up and extend HlsChunkSourceTest
Half of the tests were already overriding the setUp step, so
it's cleaner to just execute the setup inline with the right
parameters in all tests via a util method.

Also adding new tests for correctly throwing playlist load
errors before and after being blocked on a playlist, but not
if loading isn't blocked. One of the tests is failing at the
moment because of Issue: androidx/media#2401.

PiperOrigin-RevId: 764152270