Fix logic to throw playlist load error when getting stuck
The current logic sets a boolean and also compares if the
loaded playlist matches the one that had a playlist load error.
However, the logic is a bit convoluted and relies on the error
being reported after getting stuck on the playlist reload.

This doesn't work if the loader already gave up (e.g. because
the load error handling policy doesn't allow further retries).

Issue: androidx/media#2401
PiperOrigin-RevId: 764207757