Use published duration of preload parts in getNextChunk
Preload parts don't have a duration initially and startTime==endTime
in the HlsMediaChunk. However, we get a proper duration later once
the playlist has been updated again.

If the last loaded chunk was a preload part, we currently pass in
its wrong end time to getNextChunk (potentially selecting a new chunk
at an earlier time than we have to), and also use the wrong end time
to decide if the chunk needs to be spliced in.

This can be fixed by retrieving the updated duration at the point
of marking the chunk as published, and then using this duration when
calculating the loadPositionUs for getNextChunk. And also use the
same loadPositonUs when deciding if the new chunk needs to be spliced
in.

PiperOrigin-RevId: 764279432