Change PlayerWrapper's default for volume without command
The command check is executed before bundling the volume into PlayerInfo. A zero will be bundled because the check against 1f fails in toBundleForRemoteProcess. Therefore we always end up passing the zero for the field we shouldn't even be accessing because we don't have the correct Player.Command.

Making the volume check return 1f brings the two systems in sync and also returns a more reasonable default (it's unlikely the Player is muted).

PiperOrigin-RevId: 764286512