Avoid splicing into HLS chunks if the splice point is already read
When splicing data into existing SampleQueues, we need to find the
first keyframe in the new data after the current reading position.
If we already read too much data and the new chunk starts before our
current reading position, the splice will fail and only the data from
the following chunk will be written to the queue, causing a gap in
the data where playback will appear stuck.

To avoid this issue mostly, we can check if the new chunk starts from
a position before the current reading position, and if so, we can
ignore the new track selection and continue loading the current format
instead to avoid the gap.

PiperOrigin-RevId: 764650636