Fix connection race condition in MediaControllerImplLegacy
When connecting to a framework or media1 controller, we ask the
session for its "extra" binder that is used to exchange information.

Once this is done we get a callback and mark the controller as
connected. For framework-only sessions however, we won't get a
callback. The only way to know is to wait for a bit and see if
any callback (onSessionReady/onSessionDestroyed) arrives.

At the moment, the only waiting is sending a handler message to the
main thread to accept the controller as a framework controller.
In about 70% of cases in the tests, this triggers the connection
too early before the onSessionReady callback comes back. We can fix
this by using a longer wait (same timeout as in SessionToken for
a similar problem). Since raw platform MediaSession instances are
rare, this is not expected to influence many practical scenarios.

Added a new test to confirm that connecting a new controller to a
released session throws a SecurityException as explained in the
Javadoc. This test fails in 70% of the time without this fix as
the session is falsely marked as connected even though it's already
released.

PiperOrigin-RevId: 765238374