Allow enabled BufferingVideoSinks to feed InputVideoSink when ready
When prewarming video renderers, there is a period where the first media item is fully input into the VideoGraph but not fully played yet, and the second renderer is enabled but not yet started. This change means we will now activate the second BufferingVideoSink on the secondary renderer at this point and start feeding frames from the second media item into the VideoGraph.

PiperOrigin-RevId: 766557694