Assume seeking to the end inside the queue can work if loading finished
When checking if loading is possible inside the SampleQueue, we don't
usually assume that samples beyond the largest queued timestamp are
acceptable because there are likely some missing samples that haven't
been loaded yet. However, we should be able to assume that the sample
could be found in the SampleQueue if loading already finished. For
DASH's ChunkSampleStream, we already had similar logic for assuming
that samples before the end of the current chunk are all loaded.

PiperOrigin-RevId: 767144080