Create an `IdlingMediaCodecAdapterFactory` for Robolectric tests
This automatically idles the queueing and callback threads on every ms
of the `Looper` that it is instantiated on.

This helps to de-flake playback tests using MediaCodec in async mode, by
ensuring the buffers are reliably queued and dequeued as the playback
thread advances.

The factory also provides an explicit `idleQueueingAndCallbackThreads()`
method.

PiperOrigin-RevId: 767603635