Fix DASH thumbnail decoding downscaling
We recently limited bitmap decoding to the screen size by default:
https://github.com/androidx/media/commit/80b7045efe9d3863067c37fc3bf98d91129b6abb

This inadvertantly also affected whole DASH thumbnail grid images, but
it's more correct to limit the size of each thumbnail to the screen
size.

This issue was revealed when removing the deprecated usage of
`ImageDecoder.Factory.DEFAULT` in the test-only
`CapturingRenderersFactory`, so this change also fixes that to make the
test more realistic.

PiperOrigin-RevId: 768094880