Move platform session logic from PlayerWrapper to MSLegacyStub
PlayerWrapper keeps a lot of state and helper methods for the
platform session even though the actual user of this data is
almost exclusively MediaSessionLegacyStub where the platform
session state is handled. And for the few usages outside of
this class, it becomes clearer that this refers to the platform
session if the data is accessed through MediaSessionLegacyStub.

PiperOrigin-RevId: 770128118