Use existing PlayerInfo when connecting controllers
We currently create a new instance from scratch even though we
already have one with the latest state available. This is slightly
error-prone because the method to create a PlayerInfo from scratch
is only meant to be used for the initial state where we have to
guess some values (like change reasons).

PiperOrigin-RevId: 770129353