Fix MediaController state masking for seeking in live streams
When seeking in live streams, we currently use the period position
as the window position, which may crash the controller on the long
to integer conversion in Util.percentInt when attempting to use
this position to calculate a buffered percentage.

Also increase test coverage for more complicated media setups by
introducing a period-window-offset in PlaylistTimeline (which
makes the corresponding tests in MediaControllerStateMaskingTest
fail).

PiperOrigin-RevId: 770169767