Mark `WebServerDispatcher` and nested classes as `final`
The constructors for `WebServerDispatcher` and `Resource` are already
private, so this is a no-op for these types. The constructor for
`Resource.Builder` is public, but it seems clear that subclassing this
would be unexpected, so forbidding it seems reasonable. Adding this
keyword to all the classes makes the intention to disallow subclassing
clearer and more future-proof.

PiperOrigin-RevId: 770214319