Update platform session position compat logic for live streams
For live streams, we already clear out the position from the platform
session to prevent the live window duration from being shown to users
via the platform session. However, we still publish the position within
this live window, which should also be reset. In addition, the playback
speed needs to be set to zero to prevent controller UIs from
interpolating the position forward. The duration should also not just
be cleared (as it will default to zero), but explicitly set to a
negative value according to METADATA_KEY_DURATION. Finally, if there is
no position and duration, we also need to remove the ACTION_SEEK_TO to
disable scrubber elements in the UI because seeking in the current item
isn't possible without a position and duration.

None of this affects Media3 controllers who can already check if a
stream is live and provide a better experience to the user.

Issue: androidx/media#1758
PiperOrigin-RevId: 770590067