Disable sleeping for Offload if conditions enabling sleep change
Currently we trigger sleeping for offload if sleep has been requested(underlying audio track buffers are full) and we are "playing". We are "playing" if the player is in a READY state, playWhenReady is `true`, and playback is not being suppressed. If "playing" becomes false then the player is no longer "sleeping for offload".

PiperOrigin-RevId: 770608824