Prevent lingering messages in handler when releasing ExoPlayer
This change ensure we are also removing all the messages in the player's
handler when releasing it. Previously, we just explicitly removed all the `MSG_DO_SOME_WORK` as the result of `resetInternal()`.

This change also fix the javadoc of `TestExoPlayerBuilder.setLooper`. Previously, the expression of "used by the player" easily gave an wrong impression that this will be used as the playback looper.

PiperOrigin-RevId: 771102494