Ensure PlayerSurface views can be reused in reusable containers
When the AndroidView created instances are reused, it currently
fails to assign the surface to the player correctly.

 - The local remembered `view` variable is reset and wasn't assigned
   to the already existing view because the `factory` wasn't run
   again. Instead we should assign the local variable in `update`,
   which is run at the point of reusing the view.
 - To avoid assigning the same surface to multiple players, we need
   to know which player it was assigned to previously. Since the
   remembered Compose states are all cleared, we can instead store
   the currently assigned player as a `tag` in the View itself. This
   removes the need to have the `registeredPlayer` local state as
   it now trackes the assigned player within a single usage and
   across reuses.
 - The LaunchedEffect can now be run when the view is reused, but
   the player hasn't actually changed. So we can add an extra check
   to avoid clearing the player if it hasn't changed. It's still
   useful to keep setting the same player twice as it re-triggers
   size callbacks and first frame callbacks which might be needed by
   the new usage.

Also add new test coverage to capture the usage in reusable containers
like LazyColumn, both with different players, the same player, and two
players (so that the same player reuses the same surface as before).
All 3 tests can also be done with `null` assignments to the invisible
elements to ensure that the transition via `null` doesn't cause any
issues.

To actually assert everything works as expected, we need to track all
surface interactions more closely and then check the test setup
assumptions, and the correct and effective usage via helper methods.
This is needed as we can't know exactly how many instances are reused
by the LazyColumn in the test and when they are reset, etc.

Issue: androidx/media#2493
PiperOrigin-RevId: 774743304