Avoid splice in HLS chunks if it can be fulfilled by discarding
This is an extension to https://github.com/androidx/media/commit/6c24cc1f446d9e142cc7d8cb3d52e7a0d4f97f61. While the previous change
prevents selecting a new chunk that forces an impossbile splice
operation caused by having already read too far, it doesn't prevent
the player from reading into the to-be-replaced data later before
the chunk can be spliced in. This may still cause stuck playbacks
while we are waiting for a suitable splice-in point.

Since the most common case during low-latency playback for this
situation is going back to a previous independent part or segment,
we should be able to avoid the splice entirely by discarding the
chunks we are going to replace anyway. If we do this, then we
don't need to worry reading into them later. The main conditions
that need to be fulfilled for this to work are:
 - The new chunk must be independent (= we know the first sample
   could be spliced in successfully anyway)
 - The replaced chunk must be declared to start at the same time
   (=we don't need samples of a previous chunk to remain and can
   discard it entirely). Note that we use the declared start time
   from the playlist (not the actual sample times) so that this
   works in the same way as any other ABR switch even if the actual
   sample times are not perfectly aligned.

Issue: androidx/media#2299

#cherrypick

PiperOrigin-RevId: 775615562