Delay onMediaItemChanged() call until input is queued
This is prework for plumbing a position offset to AudioProcessor
instances via AudioGraphInput to allow seek handling.
onMediaItemChanged() needs to be delayed because the exact per MediaItem
presentation timestamp is only known after the seek is resolved by
MediaCodecAudioRenderer and passed on the first call to #handleInput().

This is a functional change, but should not change the visible order of
calls to AudioGraphInput.

PiperOrigin-RevId: 776549660