Workaround for race condition in `setSamplingGlShaderProgram`
If apps rapidly issue `flush()` and `registerInputStream()` to `DefaultVideoFrameProcessor`, this event ordering could happen

1. `ExternalTextureManager.setExternalGlShaderProgram()` is called, when reconfiguring the pipeline
1. `ExternalTextureManager.signalEndOfCurrentInputStream()` is called again, in response to prepare for the next reconfiguration (via `DefaultVideoFrameProcessor.registerInputStream()`

However, because `setExternalGlShaderProgram()` posts the actual setting of the shader program to run later, when (2) executes, it's possible the `ExternalTextureManager` hasn't set the new shader program just yet and the EOS is routed to the old one.

The change in `setSamplingGlShaderProgram` is to combat this issue. The other issue it solves at the same time is, it makes sure that the operation of setting the new shader program could not be flushed (because it's not posted to the task executor anymore)

**The change above introduces another race-y issue**

Imagine an event order

1. `onReadyToAcceptInputFrame()` is called on `SP1`
1. `setExternalGlShaderProgram()` is called to update to `SP2`
1. `onReadyToAcceptInputFrame()` is called again on `SP2`

Previously, as all method calls post to the executor, `SP1` and `SP2` capacities are updated correctly. But now as (2) runs directly, (the events are called in the same order but) could be executed as

1. `onReadyToAcceptInputFrame()` is called, but **runs 2nd**
1. `setExternalGlShaderProgram()` is called, but **runs 1st**
1. `onReadyToAcceptInputFrame()` is called again, but **runs 3rd**

And because `onReadyToAcceptInputFrame()` doesn't distinguish which shader program should receive the capacity increment, `SP2` receives `onReadyToAcceptInputFrame()` twice, thinking it can receive two frames but in reality it can receive only one.

PiperOrigin-RevId: 776598859