[ui-compose] Add `ProgressStateWithTickInterval` for position/duration
This ProgressState stateholder is customized to be used by text-based UI elements that are ticking with the media-time. The "grid" for ticking is chosen as the multiples of tickInterval (e.g. 1s, 2s, 3s) and the `position` is only updated then.

Text-based UI element is likely to be a string, whose granularity (e.g. for HH:MM:SS seconds are the highest granularity) should be matched with the tickInterval to help syncronise recomposition while keeping it as low as possible. This helps avoid the responsibility of the consumer to use derivedState to cap the recompositions.

[demo-compose] Add `TextProgressIndicator`, i.e. textual representations of progress, to `demo-compose`.

PiperOrigin-RevId: 777567494