Make `FakeMediaPeriod.seekToUs` more realistic
This change passes `allowTimeBeyondBuffer = isLoadingFinished()`,
[similar to `ProgressiveMediaPeriod`](https://github.com/androidx/media/blob/5a14369938970909b258d61c591ed182b20721a4/libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/ProgressiveMediaPeriod.java#L986).

This causes `DefaultAnalyticsCollectorTest.dynamicTimelineChange` to
fail because it was expecting is `isLoading` to toggle twice, once in
the first period and once in the second period. This was happening
before because the `moveMediaItems` call results in seeking to
`132900000`, which is beyond the end of the first item's sample stream,
which fails with `allowTimeBeyondBuffer = false`, causing the
`SampleStream` to be reset, and the data to be loaded again.

After this change, the seek succeeds and the sample stream isn't
reset, and since loading is instantaneous it all happens during the
first period, so this change updates the test to reflect this.

PiperOrigin-RevId: 778084324