Ensure every controller call is eventually finished with new PlayerInfo
When controllers send commands, they start masking their state until
all these commands have been acknowledged and a new PlayerInfo arrived
after they've been acknowledged. This usually works because all these
calls change the PlayerInfo in the session in some way.

However, if the PlayerInfo stays unchanged, no such update is sent and
the controller stays in a masking state forever. To fix this, we can
ensure a PlayerInfo is sent by scheduling one when acknowledging a
command. This should not trigger additional messages in the normal case
as we would trigger a PlayerInfo update anyway.

For backwards-compatibility we can fix this issue is a more fuzzy way
on the controller side by assuming after 500ms that no real update
from the session is arriving and ending the masking state manually.

In the masking state tests, this actually highlights a source of
flakiness caused by receiving such updates in test listeners even
though the listeners don't care about updates beyond the initial
masking state update. To fix this, we can only act on these callbacks
while we are still waiting for them. This completely removes the
flakiness of this class as well.

PiperOrigin-RevId: 778415541